allow docker daemon to be mounted inside workers
introduce a new element in submission, ``attach_docker``, which
defaults to False. when set to True, mount the Docker sock inside of
the worker, so it can run nested Docker workloads.

please note this a temporary fix and should be evolved in order
to address the obvious privacy issues of exposing the host's Docker
daemon inside of running containers.